refactor: move utils to `src\common\utils`
