+++ src/common/utils/db/errors.ts
