+++ src/province/province.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { Province, Regency } from '@prisma/client';
 import { PrismaService } from '@/prisma/prisma.service';
-import { getDBProviderFeatures } from '~/utils/db';
+import { getDBProviderFeatures } from '@/common/utils/db';
 import { SortOptions, Sorter } from '~/utils/helpers/sorter';
 
 type ProvinceSortKeys = keyof Province;
